backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "cohere>=5.20.1",
    "fastapi>=0.104.1",
    "google-generativeai>=0.8.6",
    "markdown-it-py>=4.0.0",
    "openai>=2.14.0",
    "openai-agents[litellm]>=0.6.4",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.24.0",
    "asyncpg>=0.29.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
